+++ TransactionalBox.sln
@@ -29,8 +29,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.KeyedInMemoryLock", "source\Internals\TransactionalBox.Internals.KeyedInMemoryLock\TransactionalBox.Internals.KeyedInMemoryLock.csproj", "{59EA3E37-092B-4E7C-A14F-CC6FD58709FC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.Transport.InMemory", "source\Internals\TransactionalBox.Internals.Transport.InMemory\TransactionalBox.Internals.Transport.InMemory.csproj", "{876EE663-DA64-460F-91C1-4FC281D17D00}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.DistributedLock.EntityFrameworkCore", "source\Internals\TransactionalBox.Internals.DistributedLock.EntityFrameworkCore\TransactionalBox.Internals.DistributedLock.EntityFrameworkCore.csproj", "{504D1EFF-49E7-460D-8531-ECC4B43FBBBC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.DistributedLock", "source\Internals\TransactionalBox.Internals.DistributedLock\TransactionalBox.Internals.DistributedLock.csproj", "{2EBB9634-0717-4A5C-9D83-5FC19CC15C80}"
@@ -107,10 +105,6 @@ Global
 		{59EA3E37-092B-4E7C-A14F-CC6FD58709FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{59EA3E37-092B-4E7C-A14F-CC6FD58709FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{59EA3E37-092B-4E7C-A14F-CC6FD58709FC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{876EE663-DA64-460F-91C1-4FC281D17D00}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{876EE663-DA64-460F-91C1-4FC281D17D00}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{876EE663-DA64-460F-91C1-4FC281D17D00}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{876EE663-DA64-460F-91C1-4FC281D17D00}.Release|Any CPU.Build.0 = Release|Any CPU
 		{504D1EFF-49E7-460D-8531-ECC4B43FBBBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{504D1EFF-49E7-460D-8531-ECC4B43FBBBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{504D1EFF-49E7-460D-8531-ECC4B43FBBBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -182,7 +176,6 @@ Global
 		{9C70A8B9-322F-49D5-AF08-889B6BBB161C} = {8FC46C4A-3C0C-458F-ACF5-F35F728A2988}
 		{891F6713-08CE-4FCF-881C-BBA258C18C72} = {9C70A8B9-322F-49D5-AF08-889B6BBB161C}
 		{59EA3E37-092B-4E7C-A14F-CC6FD58709FC} = {9C70A8B9-322F-49D5-AF08-889B6BBB161C}
-		{876EE663-DA64-460F-91C1-4FC281D17D00} = {9C70A8B9-322F-49D5-AF08-889B6BBB161C}
 		{504D1EFF-49E7-460D-8531-ECC4B43FBBBC} = {9C70A8B9-322F-49D5-AF08-889B6BBB161C}
 		{2EBB9634-0717-4A5C-9D83-5FC19CC15C80} = {9C70A8B9-322F-49D5-AF08-889B6BBB161C}
 		{824310B6-2605-450F-B959-88411A7BD6BE} = {6DEFBC37-33F1-4E42-9A7E-2752E9AD3187}
