--- source/Internals/TransactionalBox.Internals.Transport.InMemory/TransactionalBox.Internals.Transport.InMemory.csproj
@@ -1,12 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<ItemGroup>
-		<InternalsVisibleTo Include="TransactionalBox" />
-		<InternalsVisibleTo Include="TransactionalBox.Inbox" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
-	</ItemGroup>
-	
-</Project>
