+++ source/TransactionalBox/Internals/InternalPackages/Transport/TransportObject.cs
@@ -1,4 +1,4 @@
-﻿namespace TransactionalBox.Internals.Transport.InMemory.Internals
+﻿namespace TransactionalBox.Internals.InternalPackages.Transport
 {
     internal sealed class TransportObject
     {
