+++ source/TransactionalBox/TransactionalBox.csproj
@@ -29,6 +29,5 @@
     <ProjectReference Include="..\Internals\TransactionalBox.Internals.EventHooks\TransactionalBox.Internals.EventHooks.csproj" />
     <ProjectReference Include="..\Internals\TransactionalBox.Internals.KeyedInMemoryLock\TransactionalBox.Internals.KeyedInMemoryLock.csproj" />
     <ProjectReference Include="..\Internals\TransactionalBox.Internals.SequentialGuid\TransactionalBox.Internals.SequentialGuid.csproj" />
-    <ProjectReference Include="..\Internals\TransactionalBox.Internals.Transport.InMemory\TransactionalBox.Internals.Transport.InMemory.csproj" />
   </ItemGroup>
 </Project>
